Package refactoring. Given, When, Then spec classes were removed and step definitions were placed in more logical classes
