+++ src/main/java/com/privalia/qa/specs/SqlDatabaseGSpec.java
@@ -28,6 +28,11 @@ import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+/**
+ * Steps definitions for working with relational databases (postgresql and mysql)
+ *
+ * @author Jose Fernandez
+ */
 public class SqlDatabaseGSpec extends BaseGSpec {
