+++ src/main/java/com/privalia/qa/specs/UtilsGSpec.java
@@ -1,0 +1,383 @@
+/*
+ * Copyright (C) 2014 Stratio (http://stratio.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.privalia.qa.specs;
+
+import com.csvreader.CsvReader;
+import com.privalia.qa.utils.ThreadProperty;
+import cucumber.api.java.en.Given;
+import cucumber.api.java.en.Then;
+import cucumber.api.java.en.When;
+import io.cucumber.datatable.DataTable;
+import org.assertj.core.api.Fail;
+import org.hjson.JsonArray;
+import org.hjson.JsonValue;
+
+import java.io.*;
+import java.util.*;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+/**
+ * Steps definitions for other useful functions/operations
+ *
+ * @author Jose Fernandez
+ */
+public class UtilsGSpec extends BaseGSpec {
+
+    public static final int DEFAULT_TIMEOUT = 1000;
+
+    /**
+     * Default constructor.
+     *
+     * @param spec CommonG object
+     */
+    public UtilsGSpec(CommonG spec) {
+        this.commonspec = spec;
+    }
+
+
+    /**
+     * Wait seconds.
+     *
+     * @param seconds                   Seconds to wait
+     * @throws InterruptedException     InterruptedException
+     */
+    @When("^I wait '(.+?)' seconds?$")
+    public void idleWait(Integer seconds) throws InterruptedException {
+        Thread.sleep(seconds * DEFAULT_TIMEOUT);
+    }
+
+
+    /**
+     * Check value stored in environment variable "is|matches|is higher than|is lower than|contains|is different from" to value provided
+     *
+     * @param envVar        The env var to verify
+     * @param operation     Operation
+     * @param value         The value to match against the condition
+     * @throws Exception    Exception
+     */
+    @Then("^'(.+?)' ((is|matches|is higher than|is lower than|contains|is different from)) '(.+?)'$")
+    public void checkValue(String envVar, String operation, String value) throws Exception {
+        switch (operation.toLowerCase()) {
+            case "is":
+                assertThat(envVar).isEqualTo(value);
+                break;
+            case "matches":
+                assertThat(envVar).matches(value);
+                break;
+            case "is higher than":
+                if (envVar.matches("^-?\\d+$") && value.matches("^-?\\d+$")) {
+                    assertThat(Integer.parseInt(envVar)).isGreaterThan(Integer.parseInt(value));
+                } else {
+                    Fail.fail("A number should be provided in order to perform a valid comparison.");
+                }
+                break;
+            case "is lower than":
+                if (envVar.matches("^-?\\d+$") && value.matches("^-?\\d+$")) {
+                    assertThat(Integer.parseInt(envVar)).isLessThan(Integer.parseInt(value));
+                } else {
+                    Fail.fail("A number should be provided in order to perform a valid comparison.");
+                }
+                break;
+            case "contains":
+                assertThat(envVar).contains(value);
+                break;
+            case "is different from":
+                assertThat(envVar).isNotEqualTo(value);
+                break;
+            default:
+                Fail.fail("Not a valid comparison. Valid ones are: is | matches | is higher than | is lower than | contains | is different from");
+        }
+    }
+
+
+    /**
+     * Save value for future use.
+     *
+     * @param value  value to be saved
+     * @param envVar thread environment variable where to store the value
+     */
+    @Given("^I save \'(.+?)\' in variable \'(.+?)\'$")
+    public void saveInEnvironment(String value, String envVar) {
+        ThreadProperty.set(envVar, value);
+    }
+
+
+    /**
+     * Read csv file and store result in list of maps
+     *
+     * @param csvFile       the csv file
+     * @throws Exception    the exception
+     */
+    @When("^I read info from csv file '(.+?)'$")
+    public void readFromCSV(String csvFile) throws Exception {
+        CsvReader rows = new CsvReader(csvFile);
+
+        String[] columns = null;
+        if (rows.readRecord()) {
+            columns = rows.getValues();
+            rows.setHeaders(columns);
+        }
+
+        List<Map<String, String>> results = new ArrayList<Map<String, String>>();
+        while (rows.readRecord()) {
+            Map<String, String> row = new HashMap<String, String>();
+            for (String column : columns) {
+                row.put(column, rows.get(rows.getIndex(column)));
+            }
+            results.add(row);
+        }
+
+        rows.close();
+
+        commonspec.setResultsType("csv");
+        commonspec.setCSVResults(results);
+    }
+
+
+
+    /**
+     * Sort elements in envVar by a criteria and order.
+     *
+     * @param envVar   Environment variable to be sorted
+     * @param criteria alphabetical,...
+     * @param order    ascending or descending
+     */
+    @When("^I sort elements in '(.+?)' by '(.+?)' criteria in '(.+?)' order$")
+    public void sortElements(String envVar, String criteria, String order) {
+
+        String value = ThreadProperty.get(envVar);
+        JsonArray jsonArr = JsonValue.readHjson(value).asArray();
+
+        List<JsonValue> jsonValues = new ArrayList<JsonValue>();
+        for (int i = 0; i < jsonArr.size(); i++) {
+            jsonValues.add(jsonArr.get(i));
+        }
+
+        Comparator<JsonValue> comparator;
+        switch (criteria) {
+            case "alphabetical":
+                commonspec.getLogger().debug("Alphabetical criteria selected.");
+                comparator = new Comparator<JsonValue>() {
+                    public int compare(JsonValue json1, JsonValue json2) {
+                        int res = String.CASE_INSENSITIVE_ORDER.compare(json1.toString(), json2.toString());
+                        if (res == 0) {
+                            res = json1.toString().compareTo(json2.toString());
+                        }
+                        return res;
+                    }
+                };
+                break;
+            default:
+                commonspec.getLogger().debug("No criteria selected.");
+                comparator = null;
+        }
+
+        if ("ascending".equals(order)) {
+            Collections.sort(jsonValues, comparator);
+        } else {
+            Collections.sort(jsonValues, comparator.reversed());
+        }
+
+        ThreadProperty.set(envVar, jsonValues.toString());
+    }
+
+
+    /**
+     * Create a JSON in resources directory with given name, so for using it you've to reference it as:
+     * $(pwd)/target/test-classes/fileName
+     *
+     * @param fileName      name of the JSON file to be created
+     * @param baseData      path to file containing the schema to be used
+     * @param type          element to read from file (element should contain a json)
+     * @param modifications DataTable containing the modifications to be done to the base schema element
+     *                      <p>
+     *                      - Syntax will be:
+     *                      {@code
+     *                      | <key path> | <type of modification> | <new value> |
+     *                      }
+     *                      for DELETE/ADD/UPDATE/APPEND/PREPEND
+     *                      where:
+     *                      key path: path to the key to be modified
+     *                      type of modification: DELETE/ADD/UPDATE/APPEND/PREPEND
+     *                      new value: new value to be used
+     *                      <p>
+     *                      - Or:
+     *                      {@code
+     *                      | <key path> | <type of modification> | <new value> | <new value type> |
+     *                      }
+     *                      for REPLACE
+     *                      where:
+     *                      key path: path to the key to be modified
+     *                      type of modification: REPLACE
+     *                      new value: new value to be used
+     *                      json value type: type of the json property (array|object|number|boolean|null|n/a (for string))
+     *                      <br>
+     *                      <br>
+     *                      For example:
+     *                      <br>
+     *                      (1)
+     *                      If the element read is {"key1": "value1", "key2": {"key3": "value3"}}
+     *                      and we want to modify the value in "key3" with "new value3"
+     *                      the modification will be:
+     *                      | key2.key3 | UPDATE | "new value3" |
+     *                      being the result of the modification: {"key1": "value1", "key2": {"key3": "new value3"}}
+     *                      <br>
+     *                      (2)
+     *                      If the element read is {"key1": "value1", "key2": {"key3": "value3"}}
+     *                      and we want to replace the value in "key2" with {"key4": "value4"}
+     *                      the modification will be:
+     *                      | key2 | REPLACE | {"key4": "value4"} | object |
+     *                      being the result of the modification: {"key1": "value1", "key2": {"key4": "value4"}}
+     * @throws Exception    Exception
+     */
+    @When("^I create file '(.+?)' based on '(.+?)' as '(.+?)' with:$")
+    public void createFile(String fileName, String baseData, String type, DataTable modifications) throws Exception {
+        // Retrieve data
+        String retrievedData = commonspec.retrieveData(baseData, type);
+
+        // Modify data
+        commonspec.getLogger().debug("Modifying data {} as {}", retrievedData, type);
+        String modifiedData = commonspec.modifyData(retrievedData, type, modifications).toString();
+
+        // Create file (temporary) and set path to be accessible within test
+        File tempDirectory = new File(String.valueOf(System.getProperty("user.dir") + "/target/test-classes/"));
+        String absolutePathFile = tempDirectory.getAbsolutePath() + "/" + fileName;
+        commonspec.getLogger().debug("Creating file {} in 'target/test-classes'", absolutePathFile);
+        // Note that this Writer will delete the file if it exists
+        Writer out = new BufferedWriter(new OutputStreamWriter(new FileOutputStream(absolutePathFile), "UTF-8"));
+        try {
+            out.write(modifiedData);
+        } catch (Exception e) {
+            commonspec.getLogger().error("Custom file {} hasn't been created:\n{}", absolutePathFile, e.toString());
+        } finally {
+            out.close();
+        }
+
+        assertThat(new File(absolutePathFile).isFile());
+    }
+
+    /**
+     * Read the file passed as parameter, perform the modifications specified and save the result in the environment
+     * variable passed as parameter.
+     * @param baseData      file to read
+     * @param type          whether the info in the file is a 'json' or a simple 'string'
+     * @param envVar        name of the variable where to store the result
+     * @param modifications modifications to perform in the content of the file
+     * @throws Exception    Exception
+     */
+    @When("^I read file '(.+?)' as '(.+?)' and save it in environment variable '(.+?)' with:$")
+    public void readFileToVariable(String baseData, String type, String envVar, DataTable modifications) throws Exception {
+        // Retrieve data
+        String retrievedData = commonspec.retrieveData(baseData, type);
+
+        // Modify data
+        commonspec.getLogger().debug("Modifying data {} as {}", retrievedData, type);
+        String modifiedData = commonspec.modifyData(retrievedData, type, modifications).toString();
+
+        // Save in environment variable
+        ThreadProperty.set(envVar, modifiedData);
+    }
+
+    /**
+     * Read the file passed as parameter and save the result in the environment
+     * variable passed as parameter.
+     * @param baseData      file to read
+     * @param type          whether the info in the file is a 'json' or a simple 'string'
+     * @param envVar        name of the variable where to store the result
+     * @throws Exception    Exception
+     */
+    @When("^I read file '(.+?)' as '(.+?)' and save it in environment variable '(.+?)'$")
+    public void readFileToVariableNoDataTable(String baseData, String type, String envVar) throws Exception {
+        // Retrieve data
+        String retrievedData = commonspec.retrieveData(baseData, type);
+
+        // Save in environment variable
+        ThreadProperty.set(envVar, retrievedData);
+    }
+
+
+    /**
+     * Checks if an exception has been thrown.
+     *
+     * @param exception                 : "IS NOT" | "IS"
+     * @param foo                       parameter generated by cucumber because of the optional expression
+     * @param clazz                     the clazz
+     * @param bar                       parameter generated by cucumber because of the optional expression
+     * @param exceptionMsg              the exception msg
+     * @throws ClassNotFoundException   the class not found exception
+     */
+    @Then("^an exception '(.+?)' thrown( with class '(.+?)'( and message like '(.+?)')?)?")
+    public void assertExceptionNotThrown(String exception, String foo, String clazz, String bar, String exceptionMsg)
+            throws ClassNotFoundException {
+        List<Exception> exceptions = commonspec.getExceptions();
+        if ("IS NOT".equals(exception)) {
+            assertThat(exceptions).as("Captured exception list is not empty").isEmpty();
+        } else {
+            assertThat(exceptions).as("Captured exception list is empty").isNotEmpty();
+            Exception ex = exceptions.get(exceptions.size() - 1);
+            if ((clazz != null) && (exceptionMsg != null)) {
+                assertThat(ex.toString()).as("Unexpected last exception class").contains(clazz);
+                assertThat(ex.toString()).as("Unexpected last exception message").contains(exceptionMsg);
+
+            } else if (clazz != null) {
+                assertThat(exceptions.get(exceptions.size() - 1).getClass().getSimpleName()).as("Unexpected last exception class").isEqualTo(clazz);
+            }
+
+            commonspec.getExceptions().clear();
+        }
+    }
+
+
+    /**
+     * Checks the different results of a previous query
+     *
+     * @param expectedResults A DataTable Object with all data needed for check the results. The DataTable must contains at least 2 columns:
+     *                        a) A field column from the result
+     *                        b) Occurrences column (Integer type)
+     *                        <p>
+     *                        Example:
+     *                        |latitude| longitude|place     |occurrences|
+     *                        |12.5    |12.7      |Valencia  |1           |
+     *                        |2.5     | 2.6      |Madrid   |0           |
+     *                        |12.5    |13.7      |Sevilla   |1           |
+     *                        IMPORTANT: There no should be no existing columns
+     * @throws Exception      Exception
+     */
+    @Then("^There are results found with:$")
+    public void resultsMustBe(DataTable expectedResults) throws Exception {
+
+        String type = commonspec.getResultsType();
+        assertThat(type).isNotEqualTo("").overridingErrorMessage("It's necessary to define the result type");
+        switch (type) {
+            case "cassandra":
+                commonspec.resultsMustBeCassandra(expectedResults);
+                break;
+            case "mongo":
+                commonspec.resultsMustBeMongo(expectedResults);
+                break;
+            case "elasticsearch":
+                commonspec.resultsMustBeElasticsearch(expectedResults);
+                break;
+            case "csv":
+                commonspec.resultsMustBeCSV(expectedResults);
+                break;
+            default:
+                commonspec.getLogger().warn("default switch branch on results check");
+        }
+    }
+}
